+++ playground/server/api/test.post.ts
@@ -1,3 +1,3 @@
 export default defineEventHandler(async (event) => {
-  console.log('api test', event.path, event.context.security)
+  console.log('api test', event.path)
 })
