+++ src/runtime/nitro/plugins/01-hidePoweredBy.ts
@@ -2,7 +2,8 @@ import { defineNitroPlugin, removeResponseHeader } from '#imports'
 
 export default defineNitroPlugin((nitroApp) => {
   nitroApp.hooks.hook('beforeResponse', (event) => {
-    if (!event.node.res.headersSent) {
+    const { rules } = event.context.security
+    if (rules.hidePoweredBy && !event.node.res.headersSent) {
       removeResponseHeader(event, 'x-powered-by')
     }
   })
